+++ castle/cms/upgrades/upgrade_2_3_1.py
@@ -10,3 +10,9 @@ def upgrade(context, logger=None):
     setup = getToolByName(context, 'portal_setup')
     setup.runAllImportStepsFromProfile(PROFILE_ID)
     cookWhenChangingSettings(api.portal.get())
+
+    setup.runImportStepFromProfile(
+        'profile-collective.elasticsearch:default',
+        'plone.app.registry',
+        run_dependencies=False,
+    )