+++ HISTORY.rst
@@ -6,6 +6,18 @@ Changelog
 - Fix z-index issue with recurrence modal for events
   [robzonenet]
 
+New:
+
+- Upgrade to latest collective.elasticsearch.
+  New versions include,
+  `collective.celery=1.1.1`
+  `collective.elasticsearch=2.0.2`
+  `celery=4.2.1`
+  `billiard = 3.5.0.4`
+  `kombu = 4.2.1`
+
+Fixes:
+
 - Fix import Folder to not customize layout when text is empty
   [vangheem]